+++ app/Signup.tsx
@@ -1,4 +1,4 @@
-// SignupPage.tsx
+import Global from '@/constants/Global';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import type { NavigationProp } from '@react-navigation/native';
 import { useNavigation } from '@react-navigation/native';
@@ -13,51 +13,49 @@ import {
   Platform,
   SafeAreaView,
   ScrollView,
-  StyleSheet,
   Text,
   TextInput,
   TouchableOpacity,
-  View,
+  View
 } from 'react-native';
-
-// 다음 우편번호 컴포넌트 import
+import styles from '../styles/signupStyles';
 import DaumPostcode, { DaumPostcodeData } from './DaumPostcode';
 
-// 모의 요양원/보호센터 데이터
+// 요양원/보호센터 모의 데이터
 const CARE_CENTERS = [
-  { id: 1, name: "행복한 요양원", address: "서울특별시 강남구 테헤란로 123", type: "요양원" },
-  { id: 2, name: "건강한 노인복지센터", address: "서울특별시 서초구 서초대로 456", type: "복지센터" },
-  { id: 3, name: "편안한 주야간보호센터", address: "서울특별시 송파구 올림픽로 789", type: "주야간보호센터" },
-  { id: 4, name: "사랑의 요양원", address: "서울특별시 강동구 천호대로 101", type: "요양원" },
-  { id: 5, name: "희망 노인복지센터", address: "서울특별시 마포구 양화로 202", type: "복지센터" },
-  { id: 6, name: "미소 주야간보호센터", address: "서울특별시 영등포구 여의대로 303", type: "주야간보호센터" },
-  { id: 7, name: "햇살 요양원", address: "서울특별시 용산구 이태원로 404", type: "요양원" },
-  { id: 8, name: "푸른 노인복지센터", address: "서울특별시 종로구 종로 505", type: "복지센터" },
+  { id: 1, name: "행복한 요양원", centerStreetAddress: "서울특별시 강남구 테헤란로 123", type: "요양원", centerAddress: "01012" },
+  { id: 2, name: "건강한 노인복지센터", centerStreetAddress: "서울특별시 서초구 서초대로 456", type: "복지센터", centerAddress: "12315" },
+  { id: 3, name: "편안한 주야간보호센터", centerStreetAddress: "서울특별시 송파구 올림픽로 789", type: "주야간보호센터", centerAddress: "57465" },
+  { id: 4, name: "사랑의 요양원", centerStreetAddress: "서울특별시 강동구 천호대로 101", type: "요양원", centerAddress: "12321" },
+  { id: 5, name: "희망 노인복지센터", centerStreetAddress: "서울특별시 마포구 양화로 202", type: "복지센터", centerAddress: "54322" },
+  { id: 6, name: "미소 주야간보호센터", centerStreetAddress: "서울특별시 영등포구 여의대로 303", type: "주야간보호센터", centerAddress: "12312" },
+  { id: 7, name: "햇살 요양원", centerStreetAddress: "서울특별시 용산구 이태원로 404", type: "요양원", centerAddress: "65453" },
+  { id: 8, name: "푸른 노인복지센터", centerStreetAddress: "서울특별시 종로구 종로 505", type: "복지센터", centerAddress: "12312"},
 ];
 
 // 인터페이스 정의
 interface CareCenter {
   id: number;
   name: string;
-  address: string;
+  centerStreetAddress: string;
   type: string;
+  centerAddress: string;
 }
 
 interface FormData {
   name: string;
-  birthDate: Date | null;
-  phone: string;
   password: string;
   confirmPassword: string;
-  zipCode: string;
-  address: string;
-  detailAddress: string;
-  isElderly: boolean;
+  birth: Date | null;
+  number: string;
+  homeAddress: string;
+  serviceCode: "PAYPASS_SERVICE" | "CARE_SERVICE" | "ALL_SERVICE" | "NONE" | "";
+  homeStreetAddress: string;
+  homeStreetAddressDetail: string;
   careCenter: CareCenter | null;
-  selectedService: "PAYPASS_SERVICE" | "CARE_SERVICE" | "ALL_SERVICE" | "NONE" | "";
+  isElderly: boolean;
 }
 
-// Navigation 타입 정의
 type RootStackParamList = {
   Signup: undefined;
   SelectRole: undefined;
@@ -66,28 +64,27 @@ type RootStackParamList = {
 type SignupScreenNavigationProp = NavigationProp<RootStackParamList, 'Signup'>;
 
 const SignupPage: React.FC = () => {
-  const navigation = useNavigation<SignupScreenNavigationProp>();
   
+  const navigation = useNavigation<SignupScreenNavigationProp>();
   const [formData, setFormData] = useState<FormData>({
-    name: "",
-    birthDate: null,
-    phone: "",
-    password: "",
-    confirmPassword: "",
-    zipCode: "",
-    address: "",
-    detailAddress: "",
-    isElderly: false,
+    name: '',
+    password: '',
+    confirmPassword: '',
+    birth: null,
+    number: '',
+    homeAddress: '',
+    serviceCode: '',
+    homeStreetAddress: '',
+    homeStreetAddressDetail: '',
     careCenter: null,
-    selectedService: "",
+    isElderly: false
   });
 
-  const [showDatePicker, setShowDatePicker] = useState<boolean>(false);
-  const [isCareCenterModalOpen, setIsCareCenterModalOpen] = useState<boolean>(false);
-  // 🔧 다음 우편번호 검색 모달 상태
-  const [isPostcodeMode, setIsPostcodeMode] = useState<boolean>(false);
-  const [searchTerm, setSearchTerm] = useState<string>("");
-  const [filteredCenters, setFilteredCenters] = useState<CareCenter[]>(CARE_CENTERS);
+  const [showDatePicker, setShowDatePicker] = useState<boolean>(false); // 생년월일 검색 관련
+  const [isPostcodeMode, setIsPostcodeMode] = useState<boolean>(false); // 다음 우편번호 검색 모달 상태
+  const [isCareCenterModalOpen, setIsCareCenterModalOpen] = useState<boolean>(false); // 케어센터 데이터 관련
+  const [searchTerm, setSearchTerm] = useState<string>(""); // 센터 검색 관련 
+  const [filteredCenters, setFilteredCenters] = useState<CareCenter[]>(CARE_CENTERS); // 센터 검색 관련 
 
   // 검색어에 따라 요양원/보호센터 필터링
   useEffect(() => {
@@ -96,9 +93,11 @@ const SignupPage: React.FC = () => {
     } else {
       const filtered = CARE_CENTERS.filter(
         (center) =>
+
           center.name.includes(searchTerm) || 
-          center.address.includes(searchTerm) || 
-          center.type.includes(searchTerm)
+          center.centerStreetAddress.includes(searchTerm) || 
+          center.type.includes(searchTerm) ||
+          center.centerAddress.includes(searchTerm)
       );
       setFilteredCenters(filtered);
     }
@@ -114,16 +113,14 @@ const SignupPage: React.FC = () => {
     }));
   };
 
-  // 🔧 다음 우편번호 검색 시작
+  // 다음 우편번호 검색 시작
   const searchZipCode = (): void => {
     setIsPostcodeMode(true);
   };
 
-  // 🔧 다음 우편번호 API에서 주소 선택 처리
+  // 다음 우편번호 API에서 주소 선택 처리
   const handleDaumPostcode = (data: DaumPostcodeData): void => {
-    console.log('다음 우편번호 API 데이터:', data);
-    
-    // 주소 조합 로직 (문서의 예시를 참고)
+    // 주소 조합 로직
     let fullAddress = data.address;
     let extraAddress = '';
 
@@ -142,19 +139,12 @@ const SignupPage: React.FC = () => {
     // 폼 데이터 업데이트
     setFormData(prev => ({
       ...prev,
-      address: fullAddress,
-      zipCode: String(data.zonecode),
+      homeStreetAddress: fullAddress,
+      homeAddress: String(data.zonecode),
     }));
     
     // 모달 닫기
     setIsPostcodeMode(false);
-    
-    // 성공 알림
-    Alert.alert(
-      "✅ 주소 선택 완료", 
-      `우편번호: ${data.zonecode}\n주소: ${fullAddress}\n\n상세주소를 입력해주세요.`,
-      [{ text: "확인" }]
-    );
   };
 
   // 다음 우편번호 검색 모달 닫기
@@ -174,7 +164,7 @@ const SignupPage: React.FC = () => {
   const onDateChange = (event: any, selectedDate?: Date): void => {
     setShowDatePicker(false);
     if (selectedDate) {
-      handleInputChange("birthDate", selectedDate);
+      handleInputChange("birth", selectedDate);
     }
   };
 
@@ -183,58 +173,28 @@ const SignupPage: React.FC = () => {
     return `${date.getFullYear()}년 ${(date.getMonth() + 1).toString().padStart(2, '0')}월 ${date.getDate().toString().padStart(2, '0')}일`;
   };
 
-  const isFormValid = (): boolean => {
-    const requiredFields = [
-      formData.name,
-      formData.birthDate,
-      formData.phone,
-      formData.password,
-      formData.confirmPassword,
-      formData.zipCode,
-      formData.address,
-      formData.detailAddress,
-      formData.selectedService,
-    ];
-
-    const basicFieldsValid = requiredFields.every((field) => 
-      field && field.toString().trim() !== ""
-    );
-
-    const passwordsMatch = formData.password === formData.confirmPassword;
-
-    if (formData.isElderly) {
-      return basicFieldsValid && passwordsMatch && formData.careCenter !== null;
-    }
-
-    return basicFieldsValid && passwordsMatch;
-  };
-
   const prepareSignupData = () => {
     const signupData = {
       name: formData.name,
-      birthDate: formData.birthDate,
-      phone: formData.phone,
       password: formData.password,
-      zipCode: formData.zipCode,
-      address: formData.address,
-      detailAddress: formData.detailAddress,
-      isElderly: formData.isElderly,
-      careCenter: formData.careCenter,
-      selectedService: formData.selectedService,
+      birth: formData.birth ? formData.birth.toISOString().slice(0, 10) : null,
+      number: formData.number,
+      homeAddress: formData.homeAddress,
+      centerAddress: formData.careCenter?.centerAddress,
+      serviceCode: formData.serviceCode,
+      homeStreetAddress: formData.homeStreetAddress,
+      homeStreetAddressDetail: formData.homeStreetAddressDetail,
+      centerStreetAddress: formData.careCenter?.centerStreetAddress
     };
     return signupData;
   };
 
   const handleSubmit = async (): Promise<void> => {
-    if (!isFormValid()) {
-      Alert.alert("⚠️ 입력 오류", "모든 필수 항목을 올바르게 입력해주세요.");
-      return;
-    }
-    
-    const signupData = prepareSignupData();
-
+  const signupData = prepareSignupData();
     try {
-      const response = await axios.post(`http://192.168.0.100:8080/login/saveNewUser`, signupData);
+      console.log('유저 가입 전송 데이터: ', signupData)
+
+      const response = await axios.post(`${Global.URL}/login/newUser`, signupData);
       console.log('서버 응답:', response.data);
 
       Alert.alert(
@@ -247,10 +207,11 @@ const SignupPage: React.FC = () => {
           }
         ]
       );
-    } catch (error) {
-      console.error('회원가입 오류:', error);
-      Alert.alert("⚠️ 서버 오류", "회원가입에 실패했습니다. 다시 시도해주세요.");
-    }
+    } catch (error: any) {
+    const message = error?.response?.data?.message || "회원가입에 실패했습니다. 다시 시도해주세요.";
+    Alert.alert("회원 가입 실패", message);
+    console.error('로그인 실패 : ', error);
+  }
   };
 
   // 체크박스 컴포넌트
@@ -279,8 +240,8 @@ const SignupPage: React.FC = () => {
   }> = ({ service, label }) => (
     <CheckboxItem
       id={service}
-      checked={formData.selectedService === service}
-      onPress={() => handleInputChange("selectedService", service as FormData['selectedService'])}
+      checked={formData.serviceCode === service}
+      onPress={() => handleInputChange("serviceCode", service as FormData['serviceCode'])}
       label={label}
     />
   );
@@ -344,14 +305,14 @@ const SignupPage: React.FC = () => {
                 activeOpacity={0.7}
               >
                 <Calendar size={16} color="#6B7280" style={styles.dateIcon} />
-                <Text style={[styles.dateText, formData.birthDate && styles.dateTextSelected]}>
-                  {formatDate(formData.birthDate)}
+                <Text style={[styles.dateText, formData.birth && styles.dateTextSelected]}>
+                  {formatDate(formData.birth)}
                 </Text>
               </TouchableOpacity>
 
               {showDatePicker && (
                 <DateTimePicker
-                  value={formData.birthDate || new Date()}
+                  value={formData.birth || new Date()}
                   mode="date"
                   display="default"
                   onChange={onDateChange}
@@ -368,11 +329,11 @@ const SignupPage: React.FC = () => {
               </Text>
               <TextInput
                 style={styles.textInput}
-                placeholder="01000000000 숫자만 작성"
-                value={formData.phone}
+                placeholder="ex) 01012345678"
+                value={formData.number}
                 onChangeText={(text) => {
                   const numericText = text.replace(/[^0-9]/g, '');
-                  handleInputChange("phone", numericText);
+                  handleInputChange("number", numericText);
                 }}
                 keyboardType="number-pad"
                 placeholderTextColor="#9CA3AF"
@@ -427,7 +388,7 @@ const SignupPage: React.FC = () => {
                 <TextInput
                   style={[styles.textInput, styles.zipCodeInput]}
                   placeholder="우편번호"
-                  value={formData.zipCode}
+                  value={formData.homeAddress}
                   editable={false}
                   placeholderTextColor="#9CA3AF"
                 />
@@ -445,7 +406,7 @@ const SignupPage: React.FC = () => {
               <TextInput
                 style={[styles.textInput, { marginBottom: 12 }]}
                 placeholder="기본주소"
-                value={formData.address}
+                value={formData.homeStreetAddress}
                 editable={false}
                 placeholderTextColor="#9CA3AF"
               />
@@ -454,8 +415,8 @@ const SignupPage: React.FC = () => {
               <TextInput
                 style={styles.textInput}
                 placeholder="상세주소를 입력하세요"
-                value={formData.detailAddress}
-                onChangeText={(text) => handleInputChange("detailAddress", text)}
+                value={formData.homeStreetAddressDetail}
+                onChangeText={(text) => handleInputChange("homeStreetAddressDetail", text)}
                 placeholderTextColor="#9CA3AF"
               />
             </View>
@@ -497,7 +458,7 @@ const SignupPage: React.FC = () => {
                       <View style={styles.careCenterAddressRow}>
                         <MapPin size={12} color="#6B7280" style={styles.mapIcon} />
                         <Text style={styles.careCenterAddress}>
-                          {formData.careCenter.address}
+                          {formData.careCenter.centerStreetAddress}
                         </Text>
                       </View>
                       <TouchableOpacity
@@ -555,12 +516,11 @@ const SignupPage: React.FC = () => {
 
             {/* 회원가입 버튼 */}
             <TouchableOpacity
-              style={[styles.submitButton, isFormValid() && styles.submitButtonActive]}
+              style={[styles.submitButton, styles.submitButtonActive]}
               onPress={handleSubmit}
-              disabled={!isFormValid()}
               activeOpacity={0.8}
             >
-              <Text style={[styles.submitButtonText, isFormValid() && styles.submitButtonTextActive]}>
+              <Text style={[styles.submitButtonText, styles.submitButtonTextActive]}>
                 회원가입 완료
               </Text>
             </TouchableOpacity>
@@ -608,7 +568,7 @@ const SignupPage: React.FC = () => {
                     <View style={styles.careCenterItemAddressRow}>
                       <MapPin size={12} color="#6B7280" style={styles.mapIcon} />
                       <Text style={styles.careCenterItemAddress}>
-                        {item.address}
+                        {item.centerStreetAddress}
                       </Text>
                     </View>
                   </TouchableOpacity>
@@ -634,366 +594,5 @@ const SignupPage: React.FC = () => {
     </SafeAreaView>
   );
 };
-
-// 🎨 스타일 정의
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F9FAFB',
-  },
-  keyboardView: {
-    flex: 1,
-  },
-  scrollView: {
-    flex: 1,
-  },
-  scrollContent: {
-    paddingHorizontal: 16,
-    paddingVertical: 24,
-  },
-  header: {
-    backgroundColor: '#FFFFFF',
-    borderRadius: 12,
-    padding: 24,
-    marginBottom: 24,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.05,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  headerTitle: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#111827',
-    textAlign: 'center',
-    marginBottom: 8,
-  },
-  headerSubtitle: {
-    fontSize: 14,
-    color: '#6B7280',
-    textAlign: 'center',
-  },
-  formContainer: {
-    backgroundColor: '#FFFFFF',
-    borderRadius: 12,
-    padding: 24,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.05,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  inputGroup: {
-    marginBottom: 24,
-  },
-  label: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: '#374151',
-    marginBottom: 8,
-  },
-  required: {
-    color: '#EF4444',
-  },
-  textInput: {
-    backgroundColor: '#F9FAFB',
-    borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    fontSize: 16,
-    color: '#111827',
-  },
-  dateButton: {
-    backgroundColor: '#F9FAFB',
-    borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  dateIcon: {
-    marginRight: 8,
-  },
-  dateText: {
-    flex: 1,
-    fontSize: 16,
-    color: '#6B7280',
-  },
-  dateTextSelected: {
-    color: '#111827',
-  },
-  zipCodeRow: {
-    flexDirection: 'row',
-    marginBottom: 12,
-  },
-  zipCodeInput: {
-    flex: 1,
-    marginRight: 8,
-  },
-  searchButton: {
-    backgroundColor: '#2563EB',
-    borderWidth: 1,
-    borderColor: '#2563EB',
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  searchIcon: {
-    marginRight: 4,
-  },
-  searchButtonText: {
-    color: '#FFFFFF',
-    fontWeight: '500',
-    fontSize: 14,
-  },
-  errorText: {
-    fontSize: 12,
-    color: '#DC2626',
-    marginTop: 4,
-  },
-  successText: {
-    fontSize: 12,
-    color: '#059669',
-    marginTop: 4,
-  },
-  checkboxContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: 8,
-  },
-  checkbox: {
-    width: 20,
-    height: 20,
-    borderWidth: 2,
-    borderColor: '#D1D5DB',
-    borderRadius: 4,
-    marginRight: 12,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  checkboxChecked: {
-    backgroundColor: '#2563EB',
-    borderColor: '#2563EB',
-  },
-  checkboxLabel: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: '#374151',
-    flex: 1,
-  },
-  elderlySection: {
-    marginLeft: 32,
-    marginTop: 16,
-    paddingLeft: 16,
-    borderLeftWidth: 2,
-    borderLeftColor: '#DBEAFE',
-    backgroundColor: '#EFF6FF',
-    borderTopRightRadius: 8,
-    borderBottomRightRadius: 8,
-    padding: 16,
-  },
-  careCenterCard: {
-    backgroundColor: '#FFFFFF',
-    padding: 16,
-    borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderRadius: 8,
-    position: 'relative',
-  },
-  careCenterHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-start',
-    marginBottom: 8,
-  },
-  careCenterName: {
-    fontSize: 16,
-    fontWeight: '500',
-    color: '#111827',
-    flex: 1,
-  },
-  careCenterTypeTag: {
-    backgroundColor: '#DBEAFE',
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-    borderRadius: 4,
-  },
-  careCenterTypeText: {
-    fontSize: 12,
-    color: '#1E40AF',
-  },
-  careCenterAddressRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 8,
-  },
-  mapIcon: {
-    marginRight: 4,
-  },
-  careCenterAddress: {
-    fontSize: 14,
-    color: '#6B7280',
-    flex: 1,
-  },
-  removeButton: {
-    position: 'absolute',
-    top: 8,
-    right: 8,
-  },
-  careCenterSearchButton: {
-    backgroundColor: '#FFFFFF',
-    borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  careCenterSearchText: {
-    fontSize: 14,
-    color: '#6B7280',
-  },
-  warningText: {
-    fontSize: 12,
-    color: '#D97706',
-    marginTop: 4,
-  },
-  serviceOptions: {
-    gap: 8,
-  },
-  submitButton: {
-    backgroundColor: '#D1D5DB',
-    borderRadius: 8,
-    paddingVertical: 16,
-    paddingHorizontal: 24,
-  },
-  submitButtonActive: {
-    backgroundColor: '#2563EB',
-  },
-  submitButtonText: {
-    textAlign: 'center',
-    fontSize: 18,
-    fontWeight: '500',
-    color: '#6B7280',
-  },
-  submitButtonTextActive: {
-    color: '#FFFFFF',
-  },
-  // 다음 우편번호 검색 관련 스타일
-  postcodeHeader: {
-    backgroundColor: '#2563EB',
-    paddingHorizontal: 16,
-    paddingVertical: 16,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  postcodeTitle: {
-    fontSize: 18,
-    fontWeight: '600',
-    color: '#FFFFFF',
-  },
-  closeButton: {
-    backgroundColor: '#1D4ED8',
-    borderRadius: 20,
-    padding: 8,
-    minWidth: 36,
-    minHeight: 36,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  // 모달 관련 스타일
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalContainer: {
-    backgroundColor: '#FFFFFF',
-    borderRadius: 12,
-    margin: 16,
-    padding: 24,
-    maxHeight: 400,
-    width: '100%',
-    maxWidth: 400,
-  },
-  modalTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#111827',
-    marginBottom: 16,
-  },
-  modalSearchInput: {
-    backgroundColor: '#F9FAFB',
-    borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    marginBottom: 16,
-  },
-  careCenterItem: {
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: '#F3F4F6',
-  },
-  careCenterItemHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginBottom: 4,
-  },
-  careCenterItemName: {
-    fontSize: 16,
-    fontWeight: '500',
-    color: '#111827',
-    flex: 1,
-  },
-  careCenterItemTypeTag: {
-    backgroundColor: '#DBEAFE',
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-    borderRadius: 4,
-  },
-  careCenterItemTypeText: {
-    fontSize: 12,
-    color: '#1E40AF',
-  },
-  careCenterItemAddressRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  careCenterItemAddress: {
-    fontSize: 14,
-    color: '#6B7280',
-    flex: 1,
-  },
-  emptyText: {
-    textAlign: 'center',
-    color: '#6B7280',
-    paddingVertical: 16,
-  },
-  modalCloseButton: {
-    backgroundColor: '#E5E7EB',
-    borderRadius: 8,
-    paddingVertical: 12,
-    marginTop: 16,
-  },
-  modalCloseButtonText: {
-    textAlign: 'center',
-    color: '#374151',
-    fontWeight: '500',
-  },
-});
 
 export default SignupPage;
