+++ styles/signupStyles.ts
@@ -1,0 +1,363 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F9FAFB',
+  },
+  keyboardView: {
+    flex: 1,
+  },
+  scrollView: {
+    flex: 1,
+  },
+  scrollContent: {
+    paddingHorizontal: 16,
+    paddingVertical: 24,
+  },
+  header: {
+    backgroundColor: '#FFFFFF',
+    borderRadius: 12,
+    padding: 24,
+    marginBottom: 24,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.05,
+    shadowRadius: 4,
+    elevation: 2,
+  },
+  headerTitle: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#111827',
+    textAlign: 'center',
+    marginBottom: 8,
+  },
+  headerSubtitle: {
+    fontSize: 14,
+    color: '#6B7280',
+    textAlign: 'center',
+  },
+  formContainer: {
+    backgroundColor: '#FFFFFF',
+    borderRadius: 12,
+    padding: 24,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.05,
+    shadowRadius: 4,
+    elevation: 2,
+  },
+  inputGroup: {
+    marginBottom: 24,
+  },
+  label: {
+    fontSize: 14,
+    fontWeight: '500',
+    color: '#374151',
+    marginBottom: 8,
+  },
+  required: {
+    color: '#EF4444',
+  },
+  textInput: {
+    backgroundColor: '#F9FAFB',
+    borderWidth: 1,
+    borderColor: '#E5E7EB',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    fontSize: 16,
+    color: '#111827',
+  },
+  dateButton: {
+    backgroundColor: '#F9FAFB',
+    borderWidth: 1,
+    borderColor: '#E5E7EB',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  dateIcon: {
+    marginRight: 8,
+  },
+  dateText: {
+    flex: 1,
+    fontSize: 16,
+    color: '#6B7280',
+  },
+  dateTextSelected: {
+    color: '#111827',
+  },
+  zipCodeRow: {
+    flexDirection: 'row',
+    marginBottom: 12,
+  },
+  zipCodeInput: {
+    flex: 1,
+    marginRight: 8,
+  },
+  searchButton: {
+    backgroundColor: '#2563EB',
+    borderWidth: 1,
+    borderColor: '#2563EB',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  searchIcon: {
+    marginRight: 4,
+  },
+  searchButtonText: {
+    color: '#FFFFFF',
+    fontWeight: '500',
+    fontSize: 14,
+  },
+  errorText: {
+    fontSize: 12,
+    color: '#DC2626',
+    marginTop: 4,
+  },
+  successText: {
+    fontSize: 12,
+    color: '#059669',
+    marginTop: 4,
+  },
+  checkboxContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 8,
+  },
+  checkbox: {
+    width: 20,
+    height: 20,
+    borderWidth: 2,
+    borderColor: '#D1D5DB',
+    borderRadius: 4,
+    marginRight: 12,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  checkboxChecked: {
+    backgroundColor: '#2563EB',
+    borderColor: '#2563EB',
+  },
+  checkboxLabel: {
+    fontSize: 14,
+    fontWeight: '500',
+    color: '#374151',
+    flex: 1,
+  },
+  elderlySection: {
+    marginLeft: 32,
+    marginTop: 16,
+    paddingLeft: 16,
+    borderLeftWidth: 2,
+    borderLeftColor: '#DBEAFE',
+    backgroundColor: '#EFF6FF',
+    borderTopRightRadius: 8,
+    borderBottomRightRadius: 8,
+    padding: 16,
+  },
+  careCenterCard: {
+    backgroundColor: '#FFFFFF',
+    padding: 16,
+    borderWidth: 1,
+    borderColor: '#E5E7EB',
+    borderRadius: 8,
+    position: 'relative',
+  },
+  careCenterHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'flex-start',
+    marginBottom: 8,
+  },
+  careCenterName: {
+    fontSize: 16,
+    fontWeight: '500',
+    color: '#111827',
+    flex: 1,
+  },
+  careCenterTypeTag: {
+    backgroundColor: '#DBEAFE',
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 4,
+  },
+  careCenterTypeText: {
+    fontSize: 12,
+    color: '#1E40AF',
+  },
+  careCenterAddressRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 8,
+  },
+  mapIcon: {
+    marginRight: 4,
+  },
+  careCenterAddress: {
+    fontSize: 14,
+    color: '#6B7280',
+    flex: 1,
+  },
+  removeButton: {
+    position: 'absolute',
+    top: 8,
+    right: 8,
+  },
+  careCenterSearchButton: {
+    backgroundColor: '#FFFFFF',
+    borderWidth: 1,
+    borderColor: '#E5E7EB',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  careCenterSearchText: {
+    fontSize: 14,
+    color: '#6B7280',
+  },
+  warningText: {
+    fontSize: 12,
+    color: '#D97706',
+    marginTop: 4,
+  },
+  serviceOptions: {
+    gap: 8,
+  },
+  submitButton: {
+    backgroundColor: '#D1D5DB',
+    borderRadius: 8,
+    paddingVertical: 16,
+    paddingHorizontal: 24,
+  },
+  submitButtonActive: {
+    backgroundColor: '#2563EB',
+  },
+  submitButtonText: {
+    textAlign: 'center',
+    fontSize: 18,
+    fontWeight: '500',
+    color: '#6B7280',
+  },
+  submitButtonTextActive: {
+    color: '#FFFFFF',
+  },
+  // 다음 우편번호 검색 관련 스타일
+  postcodeHeader: {
+    backgroundColor: '#2563EB',
+    paddingHorizontal: 16,
+    paddingVertical: 16,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  postcodeTitle: {
+    fontSize: 18,
+    fontWeight: '600',
+    color: '#FFFFFF',
+  },
+  closeButton: {
+    backgroundColor: '#1D4ED8',
+    borderRadius: 20,
+    padding: 8,
+    minWidth: 36,
+    minHeight: 36,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  // 모달 관련 스타일
+  modalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  modalContainer: {
+    backgroundColor: '#FFFFFF',
+    borderRadius: 12,
+    margin: 16,
+    padding: 24,
+    maxHeight: 400,
+    width: '100%',
+    maxWidth: 400,
+  },
+  modalTitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: '#111827',
+    marginBottom: 16,
+  },
+  modalSearchInput: {
+    backgroundColor: '#F9FAFB',
+    borderWidth: 1,
+    borderColor: '#E5E7EB',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    marginBottom: 16,
+  },
+  careCenterItem: {
+    paddingVertical: 12,
+    borderBottomWidth: 1,
+    borderBottomColor: '#F3F4F6',
+  },
+  careCenterItemHeader: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    marginBottom: 4,
+  },
+  careCenterItemName: {
+    fontSize: 16,
+    fontWeight: '500',
+    color: '#111827',
+    flex: 1,
+  },
+  careCenterItemTypeTag: {
+    backgroundColor: '#DBEAFE',
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 4,
+  },
+  careCenterItemTypeText: {
+    fontSize: 12,
+    color: '#1E40AF',
+  },
+  careCenterItemAddressRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  careCenterItemAddress: {
+    fontSize: 14,
+    color: '#6B7280',
+    flex: 1,
+  },
+  emptyText: {
+    textAlign: 'center',
+    color: '#6B7280',
+    paddingVertical: 16,
+  },
+  modalCloseButton: {
+    backgroundColor: '#E5E7EB',
+    borderRadius: 8,
+    paddingVertical: 12,
+    marginTop: 16,
+  },
+  modalCloseButtonText: {
+    textAlign: 'center',
+    color: '#374151',
+    fontWeight: '500',
+  },
+});
+
+export default styles;
